re-organize directories; add modbus_tcp rtc processing; fix RIGOL-DG100Z channel2 command;

+++ CMSIS/Device/ST/STM32G0xx/Include/system_stm32g0xx.h
@@ -55,8 +55,8 @@
   */
 extern uint32_t SystemCoreClock;         /*!< System Clock Frequency (Core Clock) */
 
-extern const uint32_t AHBPrescTable[16];  /*!<  AHB prescalers table values */
-extern const uint32_t APBPrescTable[8];   /*!< APB prescalers table values */
+extern const uint8_t AHBPrescTable[16];  /*!<  AHB prescalers table values */
+extern const uint8_t APBPrescTable[8];   /*!< APB prescalers table values */
 
 /**
   * @}
